Use yard-solargraph plugin (#759)
* Use the yard-solargraph gem

* Remove stale/redundant tags

* Replace yieldparam_single_parameter with generic<Elem>

* Redundant code

* Remove param_tuple tag

* Remove domain directive

* Replace yieldself with yieldreceiver
