+++ lib/solargraph/type_checker.rb
@@ -468,8 +468,6 @@ module Solargraph
       return [] if parameters.empty? && arguments.empty?
       return [] if pin.anon_splat?
       if parameters.empty?
-        # Functions tagged param_tuple accepts two arguments (e.g., Hash#[]=)
-        return [] if pin.docstring.tag(:param_tuple) && arguments.length == 2
         return [] if arguments.length == 1 && arguments.last.links.last.is_a?(Source::Chain::BlockVariable)
         return [Problem.new(location, "Too many arguments to #{pin.path}")]
       end
