+++ lib/solargraph/yard_map.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'yard'
-require 'yard-solargraph'
+require 'solargraph/yard_tags'
 require 'rubygems/package'
 require 'set'
 
@@ -66,7 +66,7 @@ module Solargraph
     # @return [Set<String>]
     def rebindable_method_names
       @rebindable_method_names ||= pins_by_class(Pin::Method)
-        .select { |pin| pin.comments && pin.comments.include?('@yieldself') }
+        .select { |pin| pin.comments && pin.comments.include?('@yieldreceiver') }
         .map(&:name)
         .concat(['instance_eval', 'instance_exec', 'class_eval', 'class_exec', 'module_eval', 'module_exec', 'define_method'])
         .to_set
