+++ solargraph.gemspec
@@ -40,6 +40,7 @@ Gem::Specification.new do |s|
   s.add_runtime_dependency 'thor', '~> 1.0'
   s.add_runtime_dependency 'tilt', '~> 2.0'
   s.add_runtime_dependency 'yard', '~> 0.9', '>= 0.9.24'
+  s.add_runtime_dependency 'yard-solargraph', '~> 0.1'
 
   s.add_development_dependency 'pry'
   s.add_development_dependency 'public_suffix', '~> 3.1'
